orginized files using blueprint, there are left some errors to fix
