+++ app/database.py
@@ -1,0 +1,6 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+
+
