+++ routes/templates/Base.html
@@ -18,7 +18,7 @@
         <div class="container-fluid">
             <a class="navbar-brand ms-3" href="/home">
                 <img style="object-fit: cover"
-                     src="/static/img/whiskey-logo.jpg"
+                     src="/img/whiskey-logo.jpg"
                      width="115" height="65" class="d-inline-block align-text-top">
             </a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
