+++ routes/templates/home.html
@@ -3,7 +3,7 @@
 {% block content%}
 
 <header>
-    <div class="bg-image" style="background-image: url('static/img/bg.png'); height: 40vh;">
+    <div class="bg-image" style="background-image: url('bg.png'); height: 40vh;">
         <div class="container px-4 px-lg-5">
             <div class="text-center pt-5">
                 <h1 class="display-5 fw-bolder pt-5">Elevate Your Whiskey Experience</h1>
